app/backend/requirements.txt
azure-identity==1.13.0b3
Flask==2.2.2
openai==0.27.3
azure-cosmos-4.3.1
